+++ digiplan/map/choropleths.py
@@ -197,10 +197,28 @@ class ElectricityDemandChoropleth(Choropleth):  # noqa: D101
         return calculations.electricity_demand_per_municipality().sum(axis=1).to_dict()
 
 
+class ElectricityDemand2045Choropleth(Choropleth):  # noqa: D101
+    def get_values_per_feature(self) -> dict[int, float]:  # noqa: D102
+        return (
+            calculations.electricity_demand_per_municipality_2045(self.map_state["simulation_id"]).sum(axis=1).to_dict()
+        )
+
+
 class ElectricityDemandCapitaChoropleth(Choropleth):  # noqa: D101
     def get_values_per_feature(self) -> dict[int, float]:  # noqa: D102
         capita_demand = (
             calculations.calculate_capita_for_value(calculations.electricity_demand_per_municipality()).sum(axis=1)
+            * 1e6
+        )
+        return capita_demand.to_dict()
+
+
+class ElectricityDemandCapita2045Choropleth(Choropleth):  # noqa: D101
+    def get_values_per_feature(self) -> dict[int, float]:  # noqa: D102
+        capita_demand = (
+            calculations.calculate_capita_for_value(
+                calculations.electricity_demand_per_municipality_2045(self.map_state["simulation_id"]),
+            ).sum(axis=1)
             * 1e6
         )
         return capita_demand.to_dict()
@@ -250,7 +268,9 @@ CHOROPLETHS: dict[str, Union[Callable, type(Choropleth)]] = {
     "wind_turbines_square_statusquo": WindTurbinesSquareChoropleth,
     "wind_turbines_square_2045": WindTurbinesSquare2045Choropleth,
     "electricity_demand_statusquo": ElectricityDemandChoropleth,
+    "electricity_demand_2045": ElectricityDemand2045Choropleth,
     "electricity_demand_capita_statusquo": ElectricityDemandCapitaChoropleth,
+    "electricity_demand_capita_2045": ElectricityDemandCapita2045Choropleth,
     "heat_demand_statusquo": HeatDemandChoropleth,
     "heat_demand_capita_statusquo": HeatDemandCapitaChoropleth,
     "batteries_statusquo": BatteriesChoropleth,